Extracted propose_new_coordinates
